DriverHistory & Reports & Routes

+++ src/main/java/com/vms/demo/entity/DriverHistoryEntity.java
@@ -4,7 +4,13 @@ import java.time.ZonedDateTime;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -25,9 +31,9 @@ public class DriverHistoryEntity {
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ssZ")
     private ZonedDateTime removedDate;
     /**
-     * number of liters per kilometer the car uses on average for this driver
+     * total number of liters the car used for this driver
      */
-    private int fuelConsumption;
+    private float fuelConsumption;
     /**
      * total cost of all maintenance jobs while this driver was assigned to the car
      */
