+++ src/main/java/com/vms/demo/repository/MessageRepository.java
@@ -20,5 +20,5 @@ public interface MessageRepository extends JpaRepository<MessageEntity, Long> {
     @Modifying
     @Transactional
     @Query(value = "UPDATE Messages msg SET msg.isRead = TRUE WHERE msg.chat_id = ?1 AND msg.authorIsAdmin = ?2", nativeQuery = true)
-    Long setReadByChatID(Long chatID, Boolean authorIsAdmin);
+    int setReadByChatID(Long chatID, Boolean authorIsAdmin);
 }
